requirements.txt
pandas>=1.5.0
openpyxl>=3.0.10
qrcode[pil]>=7.4.2
Pillow>=9.0.0
